Refactor: move MCP error formatting into mcp_error_utils.ts
Review feedback: the error-formatting block did not belong in the session
manager. Move the constants (MAX_RESPONSE_BODY_LENGTH, TRUNCATION_MARKER,
UNKNOWN_ERROR, MIN/MAX_HTTP_STATUS) and helpers (asRecord, firstString,
truncateBody, baseMessage, extractHttpDetails, formatErrorRecursive,
formatError, logTransportError) verbatim into a co-located
core/src/tools/mcp/mcp_error_utils.ts, matching the existing
code_execution_utils.ts pattern.

formatError stays the module's public surface (with logTransportError, which
the session manager assigns to transport.onerror); every other helper remains
module-private. mcp_session_manager.ts drops from 296 back to 145 lines and now
holds session-manager logic only. The test file moves to
mcp_error_utils_test.ts to mirror the source; it imports the module by relative
path since these utils are deliberately not part of the @google/adk public API.

Pure code move: no behavior change, no `any`, no eslint-disable. 52 MCP tests
pass with 100% line+branch coverage of both files.
